requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
numpy==1.24.1
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
pyTelegramBotAPI==4.9.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
